web-app/.ipynb_checkpoints/requirements-checkpoint.txt
typing
requests

google-api-core
google-cloud-aiplatform
google-cloud-enterpriseknowledgegraph
google-cloud-discoveryengine

Flask
gunicorn
